drivers/power/supply: Add leardal power supplies
Linkbox2 plus and Simpad2 plus

Fixes for version 6.6 of Linux:

Use the modern gpiod, then we don't have to check whether pins a active
high or low since that is automatically handled for us.
Update Kconfig and Makefile for 6.6 support.

linkbox-plus-psy: make clr-status pin optional since it is not available
in all platforms, i.e. in linkbox2p and simpad2p it is controlled by a
pgio-hog node instead of the plus-psy.

simpad-plus-psy: remove `-gpios` from startup gpio name, it is implicit
in kernel version 6.6.x.
Signed-off-by: Eino Juhani Oltedal <ejo@datarespons.no>