sound/soc/fsl/fsl_sai: Add clock provider support
fsl_sai has clock gating feature when is sources master
clock to a pin. Integrate this into the clock framework by adding a clock
provider for this gate when master is selected.
Provide TX/RX reset on PM resume.

clean up PM framework:

The misplaced setting of BIT 30 in control register results
in channel wrapping on first boot.
Clean up PM and enable PM _prior_ to announcing clocks
Do not perform reset in resume when providing clock.

Added in ofs definition for the extra line added by imx:
+	if (sai->soc_data->mclk_with_tere && sai->mclk_direction_output)

    tree 9da7a1e787731b8eaec00f7111fb8a0f6e3296f3
    parent 101b23830d3c26e9549274d16e8d4542c8bce4af
    author Shengjiu Wang <shengjiu.wang@nxp.com> Fri May 5 15:55:22 2023 +0800
    committer Mark Brown <broonie@kernel.org> Mon May 8 08:48:53 2023 +0900

Fixed several issues with this commit for it to work with 6.6 Linux

Signed-off-by: Eino Juhani Oltedal <ejo@datarespons.no>