lm-imx-8960.c: Fixes for kernel panic on CPUM+1
commit 6e3b5380dd1fb8c82d7a3a935755e8e8b35d7a8f wm8960: Add better error
handling introduced kernel panic om CPUM+1. 6e3b5380dd1f was a
conversion of patch 0031-Better-error-handling-on-wm8960 from
karsten.sperling.opdal@laerdal.com:

https://github.com/akkodis-edge/simpad2-aen-bsp/pull/141/files#diff-3366943eb0806b29ce123e730bde8f2d7617f834de2c2eac11e883cfeab68f1a

The patch was for Linux 5.15, this commit fixes issues with the
conversion of that patch to Linux 6.6.x. The line of code causing the
kernel panic on CPUM+1 was the following:

    put_device:
           put_device(&pdev->dev);

Why put_device causes a kernel panic on reboot I do not know, the commit
mentioned above  does not explain why put_device is added other than it
is supposed to make the probing "faster and more secure". Removing the
put_device resolves the issue.